replace interpolateSerializableObject with recursivelyReplaceStrings
