+++ app/domains/flows.server.ts
@@ -7,7 +7,7 @@ import { geldEinklagenFormular } from "~/domains/geldEinklagen/formular";
 import { geldEinklagenVorabcheck } from "~/domains/geldEinklagen/vorabcheck";
 import type { Config } from "~/services/flow/server/buildFlowController";
 import type { FlowTransitionConfig } from "~/services/flow/server/flowTransitionValidation";
-import type { Replacements } from "~/util/fillTemplate";
+import type { Replacements } from "~/util/recursivelyReplaceStrings";
 import type { Context } from "./contexts";
 import type { FlowId } from "./flowIds";
 import type { Guards } from "./guards.server";
