+++ app/routes/beratungshilfe.zustaendiges-gericht.auswahl.$PLZ.tsx
@@ -10,7 +10,7 @@ import Heading from "~/components/Heading";
 import RichText from "~/components/RichText";
 import { fetchMeta, fetchTranslations } from "~/services/cms/index.server";
 import { edgeCaseStreets } from "~/services/gerichtsfinder/amtsgerichtData.server";
-import { fillTemplate } from "~/util/fillTemplate";
+import { recursivelyReplaceStrings } from "~/util/recursivelyReplaceStrings";
 import { splitObjectsByFirstLetter } from "~/util/strings";
 
 export const loader = async ({ params, request }: LoaderFunctionArgs) => {
@@ -30,10 +30,10 @@ export const loader = async ({ params, request }: LoaderFunctionArgs) => {
     fetchMeta({ filterValue }),
   ]);
 
-  const resultListHeading = fillTemplate({
-    template: common.resultListHeading,
-    replacements: { postcode: zipCode ?? "" },
-  });
+  const resultListHeading = recursivelyReplaceStrings(
+    common.resultListHeading,
+    { postcode: zipCode ?? "" },
+  );
 
   return {
     resultListHeading,
